Move scheme to a subproject, load dynamically
In our goal to move away from Scheme code in the c4.c file itself, this
set of changes moves the scheme runtime to an external library.

This results in a much smaller base application size, and theoratically
allows for different runtimes to be loaded.

Tested to work successfully on amd64 and or1k platforms.